--- src/screens/RecipesList/components/RecipesCards/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useRecipesCardController';
